+++ features/home/index.tsx
@@ -2,6 +2,9 @@ import { Container } from "@mui/system";
 import Carrousel from "@/components/Carrousel";
 import { useQuery } from "react-query";
 import { getSlideshow } from "services/slideshow";
+import Featured from "./featured";
+import NewListings from "./new-listing";
+import Events from "./event";
 
 const MainPage = () => {
   const { data: slideshowData, isSuccess } = useQuery(["getSlideshow"], () =>
@@ -11,6 +14,9 @@ const MainPage = () => {
   return (
     <Container sx={{ height: "100%" }} disableGutters maxWidth={false}>
       {isSuccess && <Carrousel slideshow={slideshowData} />}
+      <Featured />
+      <Events />
+      <NewListings />
     </Container>
   );
 };
